+++ Makefile
@@ -1,2 +1,2 @@
 all:
-	gcc -Isrc/include -Lsrc/lib -o main src/lib/main.c src/lib/linkedlist.c src/lib/snake.c src/lib/grid.c -lmingw32 -lSDL2main -lSDL2
+	gcc -Isrc/include -Lsrc/lib -o main src/lib/main.c src/lib/snake.c src/lib/grid.c src/lib/setup.c src/lib/scene_manager.c src/lib/graphics.c -lmingw32 -lSDL2main -lSDL2
