+++ src/include/grid.h
@@ -1,7 +1,7 @@
-#include "graphics.h"
-
 #ifndef GRID_H
 #define GRID_H
+
+#include "graphics.h"
 
 void renderGrid();
