--- src/include/position.h
@@ -1,10 +1,0 @@
-
-#ifndef POSITION_H
-#define POSITION_H
-
-typedef struct {
-    int x;
-    int y;
-} Position;
-
-#endif
