+++ src/lib/graphics.c
@@ -1,0 +1,10 @@
+#include "graphics.h"
+
+SDL_Window *gWindow = NULL;
+SDL_Surface *gScreenSurface = NULL; 
+SDL_Renderer *gRenderer = NULL;
+SDL_Surface *textSurface = NULL;
+
+SDL_Rect foodRect = {(int)NULL, (int)NULL, 16,16};
+SDL_Rect snakeRect = {(int)NULL, (int)NULL, 16,16};
+SDL_Rect bodyRect = {(int)NULL, (int)NULL, 16,16};
